requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.9
httpx==0.27.2
websockets==13.1
python-socketio==5.11.0

# GraphQL
strawberry-graphql[fastapi]==0.216.1
graphql-core==3.2.3

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.10

# Redis & Caching
redis[hiredis]==5.2.0
aiocache==0.12.3

# Task Queue
celery[redis]==5.4.0
flower==2.0.1

# ML/AI Core
torch==2.5.0
torchaudio==2.5.0
transformers==4.45.0
datasets==3.0.0
accelerate==0.34.0
optimum==1.22.0
tensorflow==2.17.0
langdetect==1.0.9
openai-whisper==20231117

# Audio Processing
librosa==0.10.2.post1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
scipy==1.14.1
resampy==0.4.3

# Model Optimization
onnx==1.16.2
onnxruntime==1.20.0
tensorrt==10.6.0
triton-client==2.52.0

# Monitoring & Logging
prometheus-client==0.21.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
structlog==24.4.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.18.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==42.0.8

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2
locust==2.29.1
black==24.8.0
flake8==7.1.1
mypy==1.11.2
pre-commit==3.8.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
click==8.1.7
rich==13.9.2
tenacity==9.0.0
python-dateutil==2.9.0
pytz==2024.2

# API Documentation
pydantic-openapi-schema==1.5.1
pydantic-settings==2.6.0